refactor codes watch video

+++ app/controllers/UserController.php
@@ -667,9 +667,13 @@ class UserController extends BaseController {
     public function removePlaylist($id){
     	$id = Crypt::decrypt($id);
     	$playlistId = Crypt::decrypt(Input::get('value'));
-    	$playlistItem = PlaylistItem::where('video_id','=',$id)
-    									->where('playlist_id','=',$playlistId)->first();
-    	$playlistItem->delete();
+    	$counter = PlaylistItem::where('video_id','=',$id)
+    									->where('playlist_id','=',$playlistId);
+    	if($counter->count()){
+	    	$playlistItem = PlaylistItem::where('video_id','=',$id)
+	    									->where('playlist_id','=',$playlistId)->first();
+	    	$playlistItem->delete();
+  		}
 
     }
     public function addToFavorites($id){
@@ -682,9 +686,13 @@ class UserController extends BaseController {
 	}
 	public function removeToFavorites($id){
 		$id = Crypt::decrypt($id);
-		$favorite = Favorite::where('user_id','=',Auth::User()->id)
-							->where('video_id','=',$id)->first();
-		$favorite->delete();					
+		$counter = Favorite::where('user_id','=',Auth::User()->id)
+							->where('video_id','=',$id);
+		if($counter->count()){
+			$favorite = Favorite::where('user_id','=',Auth::User()->id)
+								->where('video_id','=',$id)->first();
+			$favorite->delete();
+		}					
 	}
 	public function addToWatchLater($id){
 		$id = Crypt::decrypt($id);
@@ -696,9 +704,13 @@ class UserController extends BaseController {
 	}
 	public function removeToWatchLater($id){
 		$id = Crypt::decrypt($id);
-		$favorite = WatchLater::where('user_id','=',Auth::User()->id)
-							->where('video_id','=',$id)->first();
-		$favorite->delete();			
+		$counter = WatchLater::where('user_id','=',Auth::User()->id)
+							->where('video_id','=',$id);
+		if($counter->count()){					
+			$favorite = WatchLater::where('user_id','=',Auth::User()->id)
+								->where('video_id','=',$id)->first();
+			$favorite->delete();
+		}			
 	}
 	public function likeVideo($id){
 		$id = Crypt::decrypt($id);
@@ -711,9 +723,13 @@ class UserController extends BaseController {
 
 	public function unlikeVideo($id){
 		$id = Crypt::decrypt($id);
-		$unlike = Like::where('user_id','=',Auth::User()->id)
-							->where('video_id','=',$id)->first();
-		$unlike->delete();
+		$counter = Like::where('user_id','=',Auth::User()->id)
+							->where('video_id','=',$id);
+		if($counter->count()){
+			$unlike = Like::where('user_id','=',Auth::User()->id)
+								->where('video_id','=',$id)->first();
+			$unlike->delete();
+		}
 	}
 
 	public function getNotification(){
